virtual/rootapiserver: split into server and config
